removed Transaction class from models.py, adjusted imports and function calls
