+++ planetmint/__init__.py
@@ -4,7 +4,6 @@
 # Code is Apache-2.0 and docs are CC-BY-4.0
 
 from planetmint.transactions.common.transaction import Transaction  # noqa
-from planetmint import models  # noqa
 from planetmint.upsert_validator import ValidatorElection  # noqa
 from planetmint.transactions.types.elections.vote import Vote  # noqa
 from planetmint.transactions.types.elections.chain_migration_election import ChainMigrationElection
@@ -12,8 +11,8 @@ from planetmint.lib import Planetmint
 from planetmint.core import App
 
 
-Transaction.register_type(Transaction.CREATE, models.Transaction)
-Transaction.register_type(Transaction.TRANSFER, models.Transaction)
+Transaction.register_type(Transaction.CREATE, Transaction)
+Transaction.register_type(Transaction.TRANSFER, Transaction)
 Transaction.register_type(ValidatorElection.OPERATION, ValidatorElection)
 Transaction.register_type(ChainMigrationElection.OPERATION, ChainMigrationElection)
 Transaction.register_type(Vote.OPERATION, Vote)
