+++ planetmint/transactions/common/transaction.py
@@ -34,8 +34,11 @@ from planetmint.transactions.common.exceptions import (
     InvalidSignature,
     AmountError,
     AssetIdMismatch,
+    DuplicateTransaction,
 )
-from planetmint.transactions.common.utils import serialize
+from planetmint.backend.schema import validate_language_key
+from planetmint.transactions.common.schema import validate_transaction_schema
+from planetmint.transactions.common.utils import serialize, validate_txn_obj, validate_key
 from .memoize import memoize_from_dict, memoize_to_dict
 from .input import Input
 from .output import Output
@@ -81,6 +84,9 @@ class Transaction(object):
     CREATE = "CREATE"
     TRANSFER = "TRANSFER"
     ALLOWED_OPERATIONS = (CREATE, TRANSFER)
+    ASSET = "asset"
+    METADATA = "metadata"
+    DATA = "data"
     VERSION = "2.0"
 
     def __init__(
@@ -118,6 +124,8 @@ class Transaction(object):
             allowed_ops = ", ".join(self.__class__.ALLOWED_OPERATIONS)
             raise ValueError("`operation` must be one of {}".format(allowed_ops))
 
+        # TODO: Move operation specific validation to sub classes
+
         # Asset payloads for 'CREATE' operations must be None or
         # dicts holding a `data` property. Asset payloads for 'TRANSFER'
         # operations must be dicts holding an `id` property.
@@ -152,6 +160,36 @@ class Transaction(object):
         self.script = script
         self._id = hash_id
         self.tx_dict = tx_dict
+
+    def validate(self, planet, current_transactions=[]):
+        """Validate transaction spend
+        Args:
+            planet (Planetmint): an instantiated planetmint.Planetmint object.
+        Returns:
+            The transaction (Transaction) if the transaction is valid else it
+            raises an exception describing the reason why the transaction is
+            invalid.
+        Raises:
+            ValidationError: If the transaction is invalid
+        """
+        input_conditions = []
+
+        # TODO: Move to Create(Transaction)
+        if self.operation == Transaction.CREATE:
+            duplicates = any(txn for txn in current_transactions if txn.id == self.id)
+            if planet.is_committed(self.id) or duplicates:
+                raise DuplicateTransaction("transaction `{}` already exists".format(self.id))
+
+            if not self.inputs_valid(input_conditions):
+                raise InvalidSignature("Transaction signature is invalid.")
+
+        # TODO: Move to Transfer(Transaction)
+        elif self.operation == Transaction.TRANSFER:
+            self.validate_transfer_inputs(planet, current_transactions)
+
+        return self
+
+    # TODO: Make this abstract, implement unspent_outputs in sub classes
 
     @property
     def unspent_outputs(self):
@@ -529,6 +567,8 @@ class Transaction(object):
             print(f"Exception ASN1EncodeError : {e}")
             return False
 
+        # TODO: find out how to resolve this
+
         if operation == self.CREATE:
             # NOTE: In the case of a `CREATE` transaction, the
             #       output is always valid.
@@ -647,6 +687,8 @@ class Transaction(object):
         if not isinstance(transactions, list):
             transactions = [transactions]
 
+        # TODO: How to decouple this?
+
         # create a set of the transactions' asset ids
         asset_ids = {tx.id if tx.operation == tx.CREATE else tx.asset["id"] for tx in transactions}
 
@@ -690,6 +732,9 @@ class Transaction(object):
         Returns:
             :class:`~planetmint.transactions.common.transaction.Transaction`
         """
+        # NOTE: this is the place where the Transaction.resolve_class function is called. 
+        # TODO: Resolve CREATE somewhere else. Figure out where this is called with tx as something else as dict
+
         operation = tx.get("operation", Transaction.CREATE) if isinstance(tx, dict) else Transaction.CREATE
         cls = Transaction.resolve_class(operation)
 
@@ -794,6 +839,7 @@ class Transaction(object):
     def register_type(tx_type, tx_class):
         Transaction.type_registry[tx_type] = tx_class
 
+    # TODO: Figure out why this is always uses the create_txn_class 
     def resolve_class(operation):
         """For the given `tx` based on the `operation` key return its implementation class"""
 
@@ -802,7 +848,13 @@ class Transaction(object):
 
     @classmethod
     def validate_schema(cls, tx):
-        pass
+        validate_transaction_schema(tx)
+        validate_txn_obj(cls.ASSET, tx[cls.ASSET], cls.DATA, validate_key)
+        validate_txn_obj(cls.METADATA, tx, cls.METADATA, validate_key)
+        validate_language_key(tx[cls.ASSET], cls.DATA)
+        validate_language_key(tx, cls.METADATA)
+
+    # TODO: this should be in the Transfer(Transaction) class
 
     def validate_transfer_inputs(self, planet, current_transactions=[]):
         # store the inputs so that we can check if the asset ids match
