+++ tests/assets/test_digital_assets.py
@@ -33,14 +33,14 @@ def test_validate_transfer_asset_id_mismatch(b, signed_create_tx, user_pk, user_
 
 
 def test_get_asset_id_create_transaction(alice, user_pk):
-    from planetmint.models import Transaction
+    from planetmint.transactions.common.transaction import Transaction
 
     tx_create = Create.generate([alice.public_key], [([user_pk], 1)])
     assert Transaction.get_asset_id(tx_create) == tx_create.id
 
 
 def test_get_asset_id_transfer_transaction(b, signed_create_tx, user_pk):
-    from planetmint.models import Transaction
+    from planetmint.transactions.common.transaction import Transaction
 
     tx_transfer = Transfer.generate(signed_create_tx.to_inputs(), [([user_pk], 1)], signed_create_tx.id)
     asset_id = Transaction.get_asset_id(tx_transfer)
@@ -48,7 +48,7 @@ def test_get_asset_id_transfer_transaction(b, signed_create_tx, user_pk):
 
 
 def test_asset_id_mismatch(alice, user_pk):
-    from planetmint.models import Transaction
+    from planetmint.transactions.common.transaction import Transaction
     from planetmint.transactions.common.exceptions import AssetIdMismatch
 
     tx1 = Create.generate([alice.public_key], [([user_pk], 1)], metadata={"msg": random.random()})
